Resurrect tests/auto/corelib/plugin/qpluginloader/machtest
Update the CMake project file to generate the binaries that are to be
tested by tst_qpluginloader.

Update the tested architectures to arm64 and x86_64 like it was done in
2739aa98b1da164b398a37e6018c205d59debf7e for Qt 5.15.

Remove the ppcconverter script that was used to create PowerPC binaries.
This architecture is unsupported since ages.

Remove the comparison with the pointer size and the alignment check.
This isn't valid since commit 2549a88ba2a48fa2bedce97dd71a2974c6f8840a.

Remove the magic header check. This is done by the parser itself since
commit 3b49aa72fe6ec0dd0aa0c1c41fb81e874dc789fa.

Remove the blacklisting of the test.

Fixes: QTBUG-86792
Change-Id: Ib7c8e648bdacca6d6290c0f40a3bb02987365fc9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>